Add CAST expression support and fix aggregate fragment rendering
Added a new Cast SQL expression.
Fixed aggregate SQL fragment generation so child expressions are separated correctly without inserting commas between their internal fragments.
Updated Average unit tests to cast SQL Server integer values to decimal, producing consistent results across SQL Server and PostgreSQL.

+++ CarriganSqlTools/Carrigan.SqlTools.IntegrationTests/CompositeModels/AggregateResult.cs
@@ -10,7 +10,7 @@ public class AggregateResult
 {
     public char? Gender { get; set; }
     public string? Title { get; set; }
-    public int? Average { get; set; }
+    public decimal? Average { get; set; }
     public int? Max { get; set; }
     public int? Min { get; set; }
     public int? Count { get; set; }
